requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openpyxl==3.1.2